+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Publishing date not updated ([#13](https://github.com/iomega/zenodo-upload/issues/13))
+* Unable to use concept DOI ([#12](https://github.com/iomega/zenodo-upload/issues/12))
 
 ## [0.4.1] - 2020-04-01